new Commit

+++ src/components/MyOrders.js
@@ -3,6 +3,8 @@ import { useEffect } from "react";
 import { useState } from "react";
 import Moment from "react-moment";
 import {BASE_API} from "./ApiConstant";
+import swal from 'sweetalert';
+
 
 function MyOrders() {
   
@@ -12,7 +14,7 @@ function MyOrders() {
 
   useEffect(() => {
     axios
-      .get(
+      .post(
         BASE_API+"/api/orders?receiverid=" +
           sessionStorage.getItem("id")
       )
@@ -31,22 +33,43 @@ function MyOrders() {
   };
 
   const deleteDetails = (orderid) => {
-    let resp = window.confirm("Are you sure to delete this product ?");
-    if (resp) {         
-    axios.delete(BASE_API+"/api/orders/" + orderid).then((resp) => {
-      alert("Product deleted successfully");
-      axios
-      .get(
-        BASE_API+"/api/orders?receiverid=" +
-          sessionStorage.getItem("id")
-      )
-      .then((resp) => {
-        console.log(resp.data);
-        setOrders(resp.data.data);
+   // let resp = window.confirm("Are you sure to delete this product ?");
+
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this Order!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        swal("Order deleted successfully!", {
+          icon: "success",
+        });
+        axios.delete(BASE_API+"/api/orders/" + orderid).then((resp) => {
+         // alert("Product deleted successfully");
+          axios
+          .get(
+            BASE_API+"/api/orders?receiverid=" +
+              sessionStorage.getItem("id")
+          )
+          .then((resp) => {
+            console.log(resp.data);
+            setOrders(resp.data.data);
+          });
+         
       });
-     
-  });
-  }
+        
+      } else {
+        swal("Your Order is safe!",{
+          icon:"info"
+        });
+      }
+    });
+
+
+
   };
 
   return (
@@ -54,7 +77,7 @@ function MyOrders() {
       <div className="row">
         <div className="col-sm-7">
           <h4 className="p-2 ">My Purchased Orders</h4>
-          <table className="table table-bordered table-sm table-dark table-striped">
+          <table className="table table-bordered table-sm table-striped" style={{background:"#CFD8DC"}}>
             <thead>
               <tr>
                 <th>Id</th>
@@ -73,16 +96,16 @@ function MyOrders() {
                   <td>
                     <button
                       onClick={(e) => showDetails(x.orderid)}
-                      className="btn btn-primary btn-sm"
+                      className="btn btn-primary btn-md"
                     >
-                      Show Details
+                      <span className="bi bi-eye-fill"></span>
                     </button>
                     &emsp;
                     <button
                       onClick={(e) => deleteDetails(x.orderid)}
-                      className="btn btn-danger btn-sm"
+                      className="btn btn-danger btn-md"
                     >
-                      Delete
+                      <span className="bi bi-trash"></span>
                     </button>
                   </td>
                 </tr>
@@ -94,7 +117,7 @@ function MyOrders() {
           {show ? (
             <>
               <h4 className="p-2">Order Details</h4>
-              <table className="table table-bordered table-sm table-dark table-striped">
+              <table className="table table-bordered table-sm table-striped" style={{background:"#CFD8DC"}}>
                 <thead>
                   <tr>
                     <th>Id</th>
