+++ src/components/Orders.js
@@ -2,6 +2,7 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 import Moment from "react-moment";
 import {BASE_API} from "./ApiConstant";
+import swal from 'sweetalert';
 
 function Orders() {
   const [orders, setOrders] = useState([]);
@@ -10,26 +11,44 @@ function Orders() {
 
   useEffect(() => {
     axios.get(BASE_API+"/api/orders").then((resp) => {
-      console.log(resp.data);
+      console.log("Maddy",resp);
       setOrders(resp.data.data);
     });
   }, []);
 
   const deleteDetails = (orderid) => {
-    let resp = window.confirm("Are you sure to delete this product ?");
-    if (resp) {         
-    axios.delete(BASE_API+"/api/orders/" + orderid).then((resp) => {
-      alert("Product deleted successfully");
-      axios
-      .get(
-        BASE_API+"/api/orders")
-      .then((resp) => {
-        console.log(resp.data);
-        setOrders(resp.data.data);
+   // let resp = window.confirm("Are you sure to delete this product ?");
+
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this Order!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        swal("Order deleted successfully!", {
+          icon: "success",
+        });
+        axios.delete(BASE_API+"/api/orders/" + orderid).then((resp) => {
+          //alert("Product deleted successfully");
+          axios
+          .get(
+            BASE_API+"/api/orders")
+          .then((resp) => {
+            console.log(resp.data);
+            setOrders(resp.data.data);
+          });
+         
       });
-     
-  });
-  }
+
+      } else {
+        swal("Your Order is safe!",{
+          icon:"info"
+        });
+      }
+    });
   };
 
   const showDetails = (orderid) => {
@@ -43,9 +62,9 @@ function Orders() {
     <div className="container-fluid">
       <div className="row">
         <div className="col-sm-7">
-          <h4 className="p-2 text-center text-white">My Purchased Orders</h4>
-          <table className="table table-bordered table-sm table-dark table-hover table-striped">
-            <thead>
+          <h4 className="p-2 text-center text-dark">My Purchased Orders</h4>
+          <table className="table table-bordered table-sm table-light table-hover table-striped">
+            <thead className="table-dark">
               <tr>
                 <th>Id</th>
                 <th>Order Date</th>
@@ -66,16 +85,16 @@ function Orders() {
                   <td>
                   <button
                       onClick={(e) => showDetails(x.orderid)}
-                      className="btn btn-primary btn-sm"
+                      className="btn btn-primary btn-md"
                     >
-                      Show Details
+                      <span className="bi bi-eye-fill"></span>
                     </button>
                     &emsp;
                     <button
                       onClick={(e) => deleteDetails(x.orderid)}
-                      className="btn btn-danger btn-sm"
+                      className="btn btn-danger btn-md"
                     >
-                      Delete
+                      <span className="bi bi-trash"></span>
                     </button>
                   </td>
                 </tr>
@@ -87,8 +106,8 @@ function Orders() {
           {show ? (
             <>
               <h4 className="p-2">Order Details</h4>
-              <table className="table table-bordered table-dark table-hover table-striped table-sm">
-                <thead>
+              <table className="table table-bordered table-light table-hover table-striped table-sm">
+                <thead className="table-dark">
                   <tr>
                     <th>Id</th>
                     <th>Product</th>
